pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
[project]
name = "ducks-10-mans"
version = "1.4.0"
description = "A Discord bot for managing valorant 10-mans"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "attrs==25.3.0",
    "beautifulsoup4==4.13.5",
    "discord.py==2.6.1",
    "dnspython==2.7.0",
    "frozenlist==1.7.0",
    "idna==3.10",
    "multidict==6.6.4",
    "propcache==0.3.2",
    "pymongo==4.14.1",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "requests==2.32.5",
    "six==1.17.0",
    "table2ascii==1.1.3",
    "typing_extensions==4.14.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "wcwidth==0.2.13",
    "yarl==1.20.1"
]

[tool.setuptools.packages.find]
include = ["*"]
